Add files via upload
Source Code from a game about a rocket with a tether that is meant to pick up objects and place a certain amount on a platform. The amount of weight it's holding determines how much it slows down and how much power it uses. The scripts are for all the different assets about it, plus the scoring.